requirements.txt
flask
bs4
requests
chromedriver_binary-auto
selenium
beautifulsoup4
transformers
torch
fugashi
ipadic